VER3-232 refactor correlation matrix service into correlation matrix directory
