+++ src/app/study-form/parameters-repeated-measure-correlations/parameters-repeated-measure-correlations.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input, OnInit} from '@angular/core';
 import {ISUFactors} from '../../shared/ISUFactors';
 import {isNullOrUndefined} from 'util';
-import {CorrelationMatrixService} from '../../shared/correlationMatrix.service';
+import {CorrelationMatrixService} from '../correlation-matrix/correlationMatrix.service';
 import {StudyService} from "../study.service";
 import {Subscription} from "rxjs/Subscription";
